ecs-on-fargate/requirements.txt
aws-cdk-lib==2.126.0
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.1
pytest==7.4.4 